信息提供

+++ database/migrations/2020_10_09_165039_create_info_provides_table.php
@@ -21,7 +21,7 @@ class CreateInfoProvidesTable extends Migration
       $table->unsignedInteger('status')->comment('0待审核1已发布2中介3已招到4面试中5不需要6未接通7电话错8态度差')->default(0);
       $table->string('info_provideable_type', 120);
       $table->unsignedBigInteger('info_provideable_id')->nullable();
-      $table->unsignedInteger('admin_user_id')->comment('信息归属人，用于查看是后台哪个账户操作的')->nullable();
+      $table->unsignedBigInteger('admin_user_id')->comment('信息归属人，用于查看是后台哪个账户操作的')->nullable();
       $table->unsignedTinyInteger('is_admin')->comment('是否管理员发布的')->default(0);
       $table->unsignedTinyInteger('is_reward')->comment('是否奖励过')->default(0);
       $table->timestamps();
